# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for the SweetRPG Game Room domain: a user's library (owned catalog volumes), wishlists, play tables, and peer-to-peer volume loans, each with per-resource visibility controls.",
        "title": "Game Room API service",
        "termsOfService": "https://pilgrimagesoftware.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://sweetrpg.com",
            "email": "admin@sweetrpg.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://mit-license.org/"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/users/{user_id}/tables": {
            "get": {
                "description": "List a user's tables, filtered to what the caller may see.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tables"
                ],
                "summary": "List tables",
                "parameters": [
                    {
                        "type": "string",
                        "example": "user-123",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/server.TableVO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new table, defaulting to private visibility.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tables"
                ],
                "summary": "Create table",
                "parameters": [
                    {
                        "type": "string",
                        "example": "user-123",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Table name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.createTableRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.TableVO"
                        }
                    },
                    "400": {
                        "description": "name missing",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/tables/{id}": {
            "get": {
                "description": "Get one table by ID, if the caller may see it.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tables"
                ],
                "summary": "Get table",
                "parameters": [
                    {
                        "type": "string",
                        "example": "user-123",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "tbl-789",
                        "description": "Table ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.TableVO"
                        }
                    },
                    "404": {
                        "description": "no table with this ID exists, or the caller may not see it",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    }
                }
            },
            "put": {
                "description": "Change a table's name. Owner-only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tables"
                ],
                "summary": "Rename table",
                "parameters": [
                    {
                        "type": "string",
                        "example": "user-123",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "tbl-789",
                        "description": "Table ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.createTableRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.TableVO"
                        }
                    },
                    "400": {
                        "description": "name missing",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    },
                    "403": {
                        "description": "table exists but is owned by a different user",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    },
                    "404": {
                        "description": "no table with this ID exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    }
                }
            },
            "delete": {
                "description": "Permanently remove a table. Owner-only.",
                "tags": [
                    "tables"
                ],
                "summary": "Delete table",
                "parameters": [
                    {
                        "type": "string",
                        "example": "user-123",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "tbl-789",
                        "description": "Table ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "table exists but is owned by a different user",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    },
                    "404": {
                        "description": "no table with this ID exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/tables/{id}/visibility": {
            "put": {
                "description": "Set a table's visibility. Owner-only. Valid visibility values: public, friends, friends_of_friends, private.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tables"
                ],
                "summary": "Set table visibility",
                "parameters": [
                    {
                        "type": "string",
                        "example": "user-123",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "tbl-789",
                        "description": "Table ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New visibility",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.visibilityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.TableVO"
                        }
                    },
                    "400": {
                        "description": "invalid body or invalid visibility value",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    },
                    "403": {
                        "description": "table exists but is owned by a different user",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    },
                    "404": {
                        "description": "no table with this ID exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/tables/{id}/volumes": {
            "post": {
                "description": "Add a catalog volume to a table's shared pool. Owner-only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tables"
                ],
                "summary": "Add table volume",
                "parameters": [
                    {
                        "type": "string",
                        "example": "user-123",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "tbl-789",
                        "description": "Table ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Volume to add",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.volumeEntryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.TableVO"
                        }
                    },
                    "400": {
                        "description": "volume_id missing",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    },
                    "403": {
                        "description": "table exists but is owned by a different user",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    },
                    "404": {
                        "description": "no table with this ID exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    }
                }
            }
        },
        "/users/{user_id}/tables/{id}/volumes/{volume_id}": {
            "delete": {
                "description": "Remove a catalog volume from a table's shared pool. Owner-only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tables"
                ],
                "summary": "Remove table volume",
                "parameters": [
                    {
                        "type": "string",
                        "example": "user-123",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "tbl-789",
                        "description": "Table ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "vol-123",
                        "description": "Volume ID",
                        "name": "volume_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.TableVO"
                        }
                    },
                    "403": {
                        "description": "table exists but is owned by a different user",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    },
                    "404": {
                        "description": "no table with this ID exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/vo.ErrorVO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "server.TableVO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "deleted_by": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "visibility": {
                    "type": "string"
                },
                "volume_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "volume_titles": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "server.createTableRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Friday night group"
                }
            }
        },
...
        "server.visibilityRequest": {
            "type": "object",
            "properties": {
                "visibility": {
                    "type": "string",
                    "example": "friends"
                }
            }
        },
        "server.volumeEntryRequest": {
            "type": "object",
            "properties": {
                "volume_id": {
                    "type": "string",
                    "example": "vol-123"
                },
                "volume_title": {
                    "type": "string",
                    "example": "Curse of Strahd"
                }
            }
        },
        "vo.ErrorVO": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}